+++ src/modules/quat/utils/quat_constants.h
@@ -10,7 +10,7 @@
 
 #define Q_PI			3.14159265f
 #define COMPASS_DECLINATION		2.36888889f//East in degrees//2.2666f//Declination +Ost-West   -10.11f	// local magnetic compass correction to true north [Cove]
-#define RAD_TO_DEG		(180.0f/Q_PI)
+#define RAD_TO_DEG		(180.0f/(float)Q_PI)
 #define DEG_TO_RAD		((float)Q_PI/180.0f)
 
 #endif /* QUAT_CONSTANTS_H_ */
